test client improvements
